+++ travel-app/package.json
@@ -30,6 +30,7 @@
     "react-images-uploading": "^3.1.2",
     "react-leaflet": "^3.1.0",
     "react-mapbox-gl": "^5.1.1",
+    "react-mapbox-gl-language": "^1.0.1",
     "react-rating-stars-component": "^2.2.0",
     "react-redux": "7.2.2",
     "react-router-dom": "5.2.0",
