--- travel-app/src/components/MapComponent/MapComponent.js
@@ -1,179 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import ReactMapboxGL, { Marker, FullscreenControl, Layer, Feature, Source, Image } from '@urbica/react-map-gl';
-import 'mapbox-gl/dist/mapbox-gl.css';
-import classes from './MapComponent.module.css';
-import { connect } from 'react-redux';
-import CountriesService from '../../services/countries-service';
-import * as actions from '../../actions/actions-country';
-import { Countries, CountriesStateType } from '../../reducers/country-reducer';
-import { RootStateType } from '../../reducers/root-reducer';
-
-const MAPBOX_ACCESS_TOKEN = 'pk.eyJ1IjoiZ2VuZXJhbC1tIiwiYSI6ImNraWozZjdrdjJkbWYycnBlNmw5N3RhNjgifQ.awd7EvjA7RM8Dl4Xb_5dBA'
-
-type MapDispatchToProps = {
-    countriesLoaded: (
-        value: Array<Countries>,
-    ) => actions.CountriesLoadedActionType;
-    countrySelect: (value: number) => actions.CountrySelectActionType;
-};
-type Props = MapDispatchToProps & CountriesStateType;
-
-const MapComponent: React.FC<Props> = ({ countriesLoaded, selectedCountryIndex, countries, selectedLanguage }) => {
-
-    let latitudeCapital = 0;
-    let longitudeCapital = 0;
-    let polygon = {};
-
-    if (countries.length > 0) {
-        latitudeCapital = countries[selectedCountryIndex].coordinate.latitude;
-        longitudeCapital = countries[selectedCountryIndex].coordinate.longitude;
-        polygon = countries[selectedCountryIndex].poligon;
-        console.log(polygon)
-    }
-
-    useEffect(() => {
-        const countryService = new CountriesService();
-        countryService.getAllCountry().then((countries) => {
-            countriesLoaded(countries.data);
-        });
-    }, [countriesLoaded]);
-
-
-    const [viewport, setViewport] = useState({
-        latitude: latitudeCapital,
-        longitude: longitudeCapital,
-        zoom: 3
-    });
-    // const layerStyle = {
-    //     id: 'anything',
-    //     type: 'fill',
-    //     paint: {
-    //         'fill-color': '#000',
-    //         'fill-opacity': 0.4
-    //     }
-    // };
-    const geojson1 = { "_id": "1", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "France", "sov_a3": "FR1", "adm0_dif": 1.0, "level": 2.0, "type": "Country", "admin": "France", "adm0_a3": "FRA", "geou_dif": 0.0, "geounit": "France", "gu_a3": "FRA", "su_dif": 0.0, "subunit": "France", "su_a3": "FRA", "brk_diff": 0.0, "name": "France", "name_long": "France", "brk_a3": "FRA", "brk_name": "France", "brk_group": null, "abbrev": "Fr.", "postal": "F", "formal_en": "French Republic", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "France", "name_alt": null, "mapcolor7": 7.0, "mapcolor8": 5.0, "mapcolor9": 9.0, "mapcolor13": 11.0, "pop_est": 64057792.0, "gdp_md_est": 2128000.0, "pop_year": -99.0, "lastcensus": -99.0, "gdp_year": -99.0, "economy": "1. Developed region: G7", "income_grp": "1. High income: OECD", "wikipedia": -99.0, "fips_10": null, "iso_a2": "FR", "iso_a3": "FRA", "iso_n3": "250", "un_a3": "250", "wb_a2": "FR", "wb_a3": "FRA", "woe_id": -99.0, "adm0_a3_is": "FRA", "adm0_a3_us": "FRA", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "Europe", "region_un": "Europe", "subregion": "Western Europe", "region_wb": "Europe & Central Asia", "name_len": 6.0, "long_len": 6.0, "abbrev_len": 3.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "MultiPolygon", "coordinates": [[[[-52.556424730018392, 2.504705308437053], [-52.939657151894977, 2.124857692875622], [-53.418465135295264, 2.053389187016037], [-53.554839240113495, 2.334896551925965], [-53.778520677288896, 2.376702785650053], [-54.088062506717279, 2.105556545414629], [-54.524754197799751, 2.311848863123785], [-54.271229620975788, 2.738747870286943], [-54.18428402364475, 3.194172268075235], [-54.011503872276819, 3.622569891774859], [-54.399542202356514, 4.212611395683481], [-54.478632981979217, 4.896755682795643], [-53.958044603070931, 5.756548163267809], [-53.618452928264844, 5.646529038918402], [-52.882141282754077, 5.409850979021599], [-51.82334286152593, 4.565768133966145], [-51.657797410678882, 4.156232408053029], [-52.249337531123984, 3.241094468596287], [-52.556424730018392, 2.504705308437053]]], [[[9.560016310269134, 42.152491970379572], [9.229752231491773, 41.380006822264448], [8.775723097375362, 41.583611965494441], [8.54421268070783, 42.256516628583086], [8.746009148807588, 42.62812185319396], [9.390000848028905, 43.009984849614739], [9.560016310269134, 42.152491970379572]]], [[[3.588184441755715, 50.378992418003577], [4.286022983425141, 49.907496649772554], [4.799221632515753, 49.985373033236328], [5.674051954784886, 49.529483547557447], [5.897759230176376, 49.442667141307169], [6.186320428094206, 49.463802802114458], [6.658229607783539, 49.201958319691556], [8.099278598674772, 49.017783515003373], [7.593676385131062, 48.333019110703731], [7.466759067422231, 47.620581976911865], [7.192202182655535, 47.449765529970989], [6.736571079138088, 47.541801255882888], [6.768713820023635, 47.287708238303679], [6.037388950228973, 46.725778713561908], [6.022609490593567, 46.272989813820516], [6.500099724970454, 46.429672756529442], [6.843592970414562, 45.991146552100673], [6.802355177445662, 45.70857982032868], [7.096652459347837, 45.333098863295874], [6.749955275101712, 45.028517971367592], [7.007562290076663, 44.254766750661389], [7.549596388386163, 44.127901109384823], [7.435184767291844, 43.693844916349178], [6.529245232783069, 43.128892320318357], [4.556962517931396, 43.399650987311588], [3.10041059735272, 43.075200507167125], [2.985998976258486, 42.473015041669896], [1.826793247087181, 42.343384711265657], [0.701590610363922, 42.795734361332649], [0.338046909190581, 42.579546006839564], [-1.502770961910471, 43.034014390630489], [-1.901351284177736, 43.422802028978339], [-1.384225226232957, 44.022610378590173], [-1.193797573237362, 46.014917710954876], [-2.225724249673789, 47.064362697938208], [-2.963276129559574, 47.570326646507965], [-4.491554938159481, 47.954954332056417], [-4.592349819344747, 48.684160468126947], [-3.295813971357745, 48.901692409859635], [-1.616510789384932, 48.644421291694584], [-1.933494025063254, 49.776341864615773], [-0.98946895995536, 49.347375800160876], [1.338761020522753, 50.127173163445264], [1.6390010921385, 50.946606350297515], [2.513573032246171, 51.148506171261857], [2.658422071960331, 50.79684804951566], [3.123251580425716, 50.780363267614518], [3.588184441755715, 50.378992418003577]]]] } }
-    const geojson0 = {
-        "_id": "0", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "Italy", "sov_a3": "ITA", "adm0_dif": 0.0, "level": 2.0, "type": "Sovereign country", "admin": "Italy", "adm0_a3": "ITA", "geou_dif": 0.0, "geounit": "Italy", "gu_a3": "ITA", "su_dif": 0.0, "subunit": "Italy", "su_a3": "ITA", "brk_diff": 0.0, "name": "Italy", "name_long": "Italy", "brk_a3": "ITA", "brk_name": "Italy", "brk_group": null, "abbrev": "Italy", "postal": "I", "formal_en": "Italian Republic", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "Italy", "name_alt": null, "mapcolor7": 6.0, "mapcolor8": 7.0, "mapcolor9": 8.0, "mapcolor13": 7.0, "pop_est": 58126212.0, "gdp_md_est": 1823000.0, "pop_year": -99.0, "lastcensus": 2012.0, "gdp_year": -99.0, "economy": "1. Developed region: G7", "income_grp": "1. High income: OECD", "wikipedia": -99.0, "fips_10": null, "iso_a2": "IT", "iso_a3": "ITA", "iso_n3": "380", "un_a3": "380", "wb_a2": "IT", "wb_a3": "ITA", "woe_id": -99.0, "adm0_a3_is": "ITA", "adm0_a3_us": "ITA", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "Europe", "region_un": "Europe", "subregion": "Southern Europe", "region_wb": "Europe & Central Asia", "name_len": 5.0, "long_len": 5.0, "abbrev_len": 5.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "MultiPolygon", "coordinates": [[[[15.520376010813834, 38.231155096991472], [15.160242954171736, 37.44404551853782], [15.309897902089006, 37.1342194687318], [15.099988234119451, 36.619987290995397], [14.335228712632016, 36.996630967754754], [13.82673261887993, 37.1045313583802], [12.431003859108813, 37.612949937483819], [12.570943637755136, 38.12638113051969], [13.741156447004585, 38.034965521795357], [14.76124922044616, 38.143873602850505], [15.520376010813834, 38.231155096991472]]], [[[9.210011834356266, 41.209991360024219], [9.809975213264977, 40.500008856766101], [9.669518670295673, 39.177376410471794], [9.214817742559489, 39.240473334300134], [8.806935662479731, 38.906617743478478], [8.428302443077115, 39.171847032216618], [8.388253208050941, 40.378310858718805], [8.15999840661766, 40.950007229163788], [8.709990675500109, 40.899984442705232], [9.210011834356266, 41.209991360024219]]], [[[12.376485223040845, 46.767559109069879], [13.806475457421556, 46.509306138691187], [13.698109978905478, 46.016778062517375], [13.937630242578336, 45.591015936864665], [13.141606479554298, 45.736691799495418], [12.328581170306308, 45.381778062514854], [12.383874952858605, 44.885374253919082], [12.261453484759159, 44.600482082694015], [12.589237094786483, 44.091365871754476], [13.526905958722494, 43.587727362637906], [14.029820997787027, 42.76100779883248], [15.142569614327954, 41.955139675456905], [15.926191033601896, 41.961315009115737], [16.169897088290412, 41.740294908203424], [15.889345737377795, 41.541082261718202], [16.785001661860576, 41.179605617836586], [17.519168735431208, 40.877143459632236], [18.376687452882578, 40.355624904942658], [18.480247023195403, 40.168866278639825], [18.293385044028099, 39.810774441073249], [17.738380161213286, 40.277671006830303], [16.869595981522338, 40.442234605463852], [16.448743116937322, 39.79540070246648], [17.171489698971499, 39.424699815420723], [17.052840610429342, 38.902871202137305], [16.635088331781844, 38.843572496082402], [16.100960727613057, 37.985898749334183], [15.684086948314501, 37.90884918878703], [15.687962680736319, 38.214592800441864], [15.891981235424709, 38.750942491199226], [16.109332309644316, 38.964547024077689], [15.718813510814641, 39.544072374014945], [15.413612501698822, 40.04835683853517], [14.998495721098237, 40.172948716790927], [14.70326826341477, 40.604550279292624], [14.060671827865264, 40.786347968095441], [13.627985060285397, 41.188287258461656], [12.88808190273042, 41.253089504555618], [12.10668257004491, 41.704534817057407], [11.191906365614187, 42.355425319989678], [10.511947869517797, 42.931462510747224], [10.200028924204048, 43.920006822274615], [9.702488234097814, 44.03627879493132], [8.888946160526871, 44.36633616797954], [8.428560825238577, 44.231228135752417], [7.850766635783202, 43.767147935555244], [7.435184767291844, 43.693844916349178], [7.549596388386163, 44.127901109384823], [7.007562290076663, 44.254766750661389], [6.749955275101712, 45.028517971367592], [7.096652459347837, 45.333098863295874], [6.802355177445662, 45.70857982032868], [6.843592970414562, 45.991146552100673], [7.273850945676685, 45.776947740250762], [7.755992058959833, 45.824490057959281], [8.316629672894379, 46.163642483090854], [8.489952426801295, 46.00515086525175], [8.966305779667834, 46.036931871111165], [9.182881707403112, 46.440214748716983], [9.922836541390353, 46.314899400409189], [10.363378126678668, 46.483571275409844], [10.442701450246602, 46.893546250997446], [11.048555942436508, 46.751358547546403], [11.164827915093326, 46.941579494812743], [12.153088006243081, 47.115393174826437], [12.376485223040845, 46.767559109069879]]]] },
-    }
-    const geojson2 = { "_id": "2", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 3.0, "sovereignt": "Thailand", "sov_a3": "THA", "adm0_dif": 0.0, "level": 2.0, "type": "Sovereign country", "admin": "Thailand", "adm0_a3": "THA", "geou_dif": 0.0, "geounit": "Thailand", "gu_a3": "THA", "su_dif": 0.0, "subunit": "Thailand", "su_a3": "THA", "brk_diff": 0.0, "name": "Thailand", "name_long": "Thailand", "brk_a3": "THA", "brk_name": "Thailand", "brk_group": null, "abbrev": "Thai.", "postal": "TH", "formal_en": "Kingdom of Thailand", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "Thailand", "name_alt": null, "mapcolor7": 3.0, "mapcolor8": 6.0, "mapcolor9": 8.0, "mapcolor13": 1.0, "pop_est": 65905410.0, "gdp_md_est": 547400.0, "pop_year": -99.0, "lastcensus": 2010.0, "gdp_year": -99.0, "economy": "5. Emerging region: G20", "income_grp": "3. Upper middle income", "wikipedia": -99.0, "fips_10": null, "iso_a2": "TH", "iso_a3": "THA", "iso_n3": "764", "un_a3": "764", "wb_a2": "TH", "wb_a3": "THA", "woe_id": -99.0, "adm0_a3_is": "THA", "adm0_a3_us": "THA", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "Asia", "region_un": "Asia", "subregion": "South-Eastern Asia", "region_wb": "East Asia & Pacific", "name_len": 8.0, "long_len": 8.0, "abbrev_len": 5.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "Polygon", "coordinates": [[[102.584932489026698, 12.186594956913282], [101.687157830819956, 12.645740057826572], [100.831809523524868, 12.627084865769206], [100.978467238369205, 13.412721665902566], [100.097797479251113, 13.406856390837433], [100.018732537844556, 12.307001044153354], [99.47892052612363, 10.846366685423547], [99.153772414143162, 9.963061428258555], [99.222398716226763, 9.239255479362427], [99.873831821698133, 9.20786204674512], [100.279646844486223, 8.295152899606052], [100.459274123132758, 7.429572658717177], [101.017327915452725, 6.856868597842478], [101.62307905477806, 6.74062246340192], [102.141186964936381, 6.221636053894628], [101.814281854257985, 5.810808417174242], [101.154218784593866, 5.691384182147715], [101.075515578213356, 6.204867051615921], [100.259596388756961, 6.642824815289543], [100.085756870527106, 6.464489447450291], [99.690690545655755, 6.848212795433597], [99.519641554769635, 7.34345388430276], [98.988252801512303, 7.907993068875328], [98.503786248775995, 8.382305202666288], [98.339661899817003, 7.794511623562386], [98.150009393305822, 8.350007432483878], [98.259150018306258, 8.973922837759801], [98.553550653073046, 9.932959906448545], [99.038120558673981, 10.960545762572437], [99.587286004639722, 11.892762762901697], [99.196353794351666, 12.80474843998867], [99.212011753336085, 13.269293728076464], [99.097755161538757, 13.827502549693278], [98.430819126379873, 14.622027696180837], [98.192074009191401, 15.123702500870351], [98.537375929765716, 15.308497422746084], [98.903348423256759, 16.177824204976119], [98.49376102091135, 16.837835598207931], [97.859122755934862, 17.567946071843664], [97.375896437573545, 18.445437730375815], [97.797782830804408, 18.627080389881755], [98.25372399291561, 19.708203029860044], [98.959675734454876, 19.752980658440947], [99.543309360759309, 20.186597601802063], [100.115987583417848, 20.417849636308187], [100.548881056726884, 20.109237982661128], [100.606293573003157, 19.508344427971224], [101.282014601651696, 19.462584947176765], [101.035931431077771, 18.408928330961615], [101.059547560635167, 17.51249725999449], [102.113591750092482, 18.109101670804165], [102.413004998791621, 17.932781683824288], [102.998705682387708, 17.961694647691601], [103.20019209189374, 18.309632066312773], [103.956476678485302, 18.240954087796879], [104.716947056092494, 17.428858954330082], [104.779320509868796, 16.441864935771449], [105.589038527450157, 15.570316066952858], [105.544338413517693, 14.723933620660418], [105.218776890078885, 14.273211778210694], [104.281418084736615, 14.416743068901367], [102.988422072361629, 14.225721136934467], [102.348099399833018, 13.394247341358223], [102.584932489026698, 12.186594956913282]]] } }
-    const geojson5 = { "_id": "5", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "Japan", "sov_a3": "JPN", "adm0_dif": 0.0, "level": 2.0, "type": "Sovereign country", "admin": "Japan", "adm0_a3": "JPN", "geou_dif": 0.0, "geounit": "Japan", "gu_a3": "JPN", "su_dif": 0.0, "subunit": "Japan", "su_a3": "JPN", "brk_diff": 0.0, "name": "Japan", "name_long": "Japan", "brk_a3": "JPN", "brk_name": "Japan", "brk_group": null, "abbrev": "Japan", "postal": "J", "formal_en": "Japan", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "Japan", "name_alt": null, "mapcolor7": 5.0, "mapcolor8": 3.0, "mapcolor9": 5.0, "mapcolor13": 4.0, "pop_est": 127078679.0, "gdp_md_est": 4329000.0, "pop_year": -99.0, "lastcensus": 2010.0, "gdp_year": -99.0, "economy": "1. Developed region: G7", "income_grp": "1. High income: OECD", "wikipedia": -99.0, "fips_10": null, "iso_a2": "JP", "iso_a3": "JPN", "iso_n3": "392", "un_a3": "392", "wb_a2": "JP", "wb_a3": "JPN", "woe_id": -99.0, "adm0_a3_is": "JPN", "adm0_a3_us": "JPN", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "Asia", "region_un": "Asia", "subregion": "Eastern Asia", "region_wb": "East Asia & Pacific", "name_len": 5.0, "long_len": 5.0, "abbrev_len": 5.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "MultiPolygon", "coordinates": [[[[134.638428176003885, 34.149233710256425], [134.766379022358507, 33.80633474378368], [134.203415968970859, 33.201177883429636], [133.792950067276507, 33.521985175097598], [133.280268182508877, 33.289570420864948], [133.014858026257883, 32.70456736910478], [132.363114862192703, 32.98938202568138], [132.371176385630207, 33.463642483040076], [132.924372593314814, 34.060298570282043], [133.492968377822223, 33.944620876596701], [133.904106073136376, 34.364931138642618], [134.638428176003885, 34.149233710256425]]], [[[140.976387567305295, 37.142074286440163], [140.599769728762141, 36.343983466124541], [140.77407433488267, 35.842877102190243], [140.253279250245129, 35.13811391859366], [138.975527785396224, 34.667600002576108], [137.217598911691226, 34.606285915661857], [135.792983026268899, 33.464805202766627], [135.120982700745429, 33.84907115328906], [135.079434849182718, 34.59654490817482], [133.340316196831992, 34.375938218720762], [132.156770868051325, 33.904933376596517], [130.98614464734348, 33.885761420216284], [132.00003624891005, 33.149992377244615], [131.332790155157369, 31.450354519164847], [130.686317987185959, 31.029579169228242], [130.202419875204981, 31.418237616495418], [130.447676222862157, 32.319474595665724], [129.814691603718899, 32.610309556604392], [129.408463169472583, 33.29605581311759], [130.353935174684665, 33.6041507024417], [130.878450962447147, 34.232742824840045], [131.88422936414392, 34.749713853487918], [132.617672967662514, 35.43339305270942], [134.608300815977799, 35.731617743465819], [135.67753787652893, 35.527134100886826], [136.723830601142453, 37.304984239240383], [137.390611607004502, 36.827390651998826], [138.857602166906275, 37.827484646143461], [139.426404657142911, 38.215962225897641], [140.054790073812086, 39.438807481436385], [139.883379347899876, 40.563312486323696], [140.30578250545372, 41.195005194659558], [141.368973423426695, 41.378559882160289], [141.914263136970504, 39.991616115878685], [141.884600864834994, 39.180864569651504], [140.959489373945786, 38.17400096287659], [140.976387567305295, 37.142074286440163]]], [[[143.910161981379503, 44.174099839853739], [144.613426548439662, 43.960882880217525], [145.320825230083102, 44.384732977875444], [145.543137241802782, 43.262088324550604], [144.059661899999895, 42.988358262700558], [143.183849725517319, 41.995214748699198], [141.611490920172486, 42.678790595056086], [141.067286411706647, 41.584593817707997], [139.955106235921079, 41.569555975911044], [139.817543573159952, 42.563758856774399], [140.312087030193226, 43.333272610032651], [141.380548944260028, 43.388824774746496], [141.67195234595394, 44.772125352551484], [141.96764489152801, 45.551483466161358], [143.142870314709825, 44.510358384776964], [143.910161981379503, 44.174099839853739]]]] } }
-    const geojson7 = { "_id": "7", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "China", "sov_a3": "CH1", "adm0_dif": 1.0, "level": 2.0, "type": "Country", "admin": "China", "adm0_a3": "CHN", "geou_dif": 0.0, "geounit": "China", "gu_a3": "CHN", "su_dif": 0.0, "subunit": "China", "su_a3": "CHN", "brk_diff": 0.0, "name": "China", "name_long": "China", "brk_a3": "CHN", "brk_name": "China", "brk_group": null, "abbrev": "China", "postal": "CN", "formal_en": "People's Republic of China", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "China", "name_alt": null, "mapcolor7": 4.0, "mapcolor8": 4.0, "mapcolor9": 4.0, "mapcolor13": 3.0, "pop_est": 1338612970.0, "gdp_md_est": 7973000.0, "pop_year": -99.0, "lastcensus": 2010.0, "gdp_year": -99.0, "economy": "3. Emerging region: BRIC", "income_grp": "3. Upper middle income", "wikipedia": -99.0, "fips_10": null, "iso_a2": "CN", "iso_a3": "CHN", "iso_n3": "156", "un_a3": "156", "wb_a2": "CN", "wb_a3": "CHN", "woe_id": -99.0, "adm0_a3_is": "CHN", "adm0_a3_us": "CHN", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "Asia", "region_un": "Asia", "subregion": "Eastern Asia", "region_wb": "East Asia & Pacific", "name_len": 5.0, "long_len": 5.0, "abbrev_len": 5.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "MultiPolygon", "coordinates": [[[[110.339187860151554, 18.678395087147607], [109.475209588663716, 18.197700913968617], [108.655207961056163, 18.507681993071401], [108.626217482540454, 19.367887885001977], [109.119055617308021, 19.821038519769388], [110.211598748822865, 20.101253973872076], [110.786550734502242, 20.077534491450081], [111.010051304164648, 19.695929877190736], [110.570646600386823, 19.255879218009312], [110.339187860151554, 18.678395087147607]]], [[[127.657407261262421, 49.760270494172943], [129.397817824420457, 49.440600084015443], [130.582293328982388, 48.729687404976119], [130.987281528853856, 47.790132351261406], [132.506671991099523, 47.788969631534883], [133.373595819228029, 48.183441677434928], [135.026311476786731, 48.478229885443909], [134.500813836810636, 47.578439846377847], [134.11236209527263, 47.212467352886733], [133.769643996312908, 46.11692698829907], [133.097126906466457, 45.144066473972174], [131.88345421765959, 45.321161607436437], [131.025212030156098, 44.967953192721581], [131.288555129115565, 44.111519680348266], [131.144687941614876, 42.929989732426947], [130.63386640840983, 42.903014634770557], [130.640015903852458, 42.395009467125277], [129.994267205933255, 42.9853868678438], [129.59666873587949, 42.424981797854599], [128.052215203972338, 41.994284572917991], [128.208433058790746, 41.466771552082548], [127.34378299368305, 41.50315176041596], [126.869083286649897, 41.816569322266162], [126.182045119329445, 41.107336127276369], [125.079941847840615, 40.569823716792456], [124.265624627785343, 39.928493353834142], [122.867570428561038, 39.637787583976262], [122.131387974130945, 39.170451768544638], [121.054554478032884, 38.897471014962917], [121.585994907722494, 39.360853583324143], [121.376757033372684, 39.750261338859531], [122.168595005381036, 40.422442531896053], [121.640358514493556, 40.946389878903318], [120.768628778161968, 40.593388169917603], [119.63960208544907, 39.898055935214217], [119.023463983233029, 39.252333075511103], [118.042748651197911, 39.204273993479688], [117.532702264477081, 38.737635809884097], [118.059698520989713, 38.061475531561058], [118.878149855628379, 37.897325344385905], [118.911636183753529, 37.448463853498737], [119.70280236214208, 37.15638865818508], [120.823457472823662, 37.870427761377982], [121.711258579597967, 37.48112335870718], [122.35793745329849, 37.454484157860691], [122.519994744965828, 36.930614325501836], [121.104163853033043, 36.65132904718044], [120.637008905114584, 36.111439520811132], [119.664561802246084, 35.609790554337735], [119.151208123858595, 34.909859117160465], [120.227524855633732, 34.36033193616862], [120.620369093916594, 33.376722723925127], [121.229014113450233, 32.460318711877193], [121.908145786630058, 31.69217438407469], [121.891919386890351, 30.949351508095106], [121.264257440273326, 30.676267401648715], [121.503519321784751, 30.142914943964257], [122.0921138855891, 29.832520453403163], [121.938428175953078, 29.01802236583481], [121.684438511238483, 28.225512600206685], [121.125661248866464, 28.135673122667185], [120.395473260582349, 27.053206895449392], [119.585496860839584, 25.740780544532612], [118.656871372554548, 24.547390855400238], [117.281606479970861, 23.624501451099718], [115.890735304835147, 22.782873236578098], [114.763827345846238, 22.66807404224167], [114.152546828265685, 22.223760077396207], [113.806779819800767, 22.54833974862143], [113.24107791550162, 22.051367499270469], [111.843592157032475, 21.550493679281516], [110.785465529424158, 21.397143866455338], [110.44403934127169, 20.341032619706397], [109.889861281373584, 20.282457383703445], [109.627655063924664, 21.008227037026728], [109.864488153118344, 21.395050970947523], [108.52281294152445, 21.715212307211829], [108.050180291782993, 21.552379869060104], [107.043420037872664, 21.811898912029903], [106.56727339073538, 22.218204860924743], [106.725403273548494, 22.794267889898379], [105.811247186305224, 22.976892401617903], [105.32920942588666, 23.352063300056983], [104.476858351664504, 22.819150092046925], [103.504514601660532, 22.70375661873922], [102.706992222100183, 22.708795070887703], [102.170435825613566, 22.464753119389343], [101.65201785686159, 22.318198757409561], [101.803119744882935, 21.174366766845054], [101.270025669360024, 21.20165192309517], [101.180005324307587, 21.43657298429406], [101.150032993578264, 21.849984442629022], [100.416537713627378, 21.558839423096657], [99.983489211021578, 21.742936713136459], [99.240898878987224, 22.118314317304566], [99.531992222087439, 22.949038804612599], [98.898749220782832, 23.142722072842588], [98.660262485755794, 24.063286037690006], [97.604719679762042, 23.897404690033056], [97.724609002679159, 25.083637193293043], [98.67183800658924, 25.918702500913497], [98.712093947344584, 26.74353587494025], [98.682690057370536, 27.508812160750665], [98.246230910233379, 27.747221381129179], [97.911987746169444, 28.335945136014374], [97.327113885490036, 28.261582749946342], [96.248833449287844, 28.411030992134471], [96.58659061074755, 28.830979519154369], [96.117678664131034, 29.45280202892252], [95.404802280664654, 29.03171662039216], [94.565990431702943, 29.277438055939967], [93.413347609432691, 28.640629380807241], [92.503118931043645, 27.896876329046449], [91.696656528696707, 27.771741848251622], [91.25885379431989, 28.04061432546635], [90.730513950567826, 28.064953925075741], [90.015828891971211, 28.296438503527185], [89.475810174521172, 28.042758897406372], [88.814248488320601, 27.299315904239393], [88.730325962278556, 28.086864732367559], [88.120440708369955, 27.876541652939579], [86.954517043000664, 27.974261786403531], [85.82331994013154, 28.203575954698749], [85.011638218123068, 28.642773952747376], [84.234579705750178, 28.839893703724698], [83.898992954446754, 29.32022614187764], [83.33711510613719, 29.463731594352197], [82.327512648450892, 30.115268052688208], [81.5258044778748, 30.422716986608663], [81.11125613802929, 30.183480943313413], [79.721366815107132, 30.882714748654735], [78.738894484374015, 31.515906073527049], [78.458446486326039, 32.618164374312727], [79.176128777995558, 32.483779812137755], [79.208891636068557, 32.994394639613745], [78.811086460285736, 33.506198025032404], [78.912268914713223, 34.321936346975775], [77.837450799474624, 35.494009507787808], [76.192848341785719, 35.898403428687857], [75.896897414050187, 36.666806138651879], [75.158027785141002, 37.133030910789159], [74.980002475895418, 37.419990139305895], [74.829985792952158, 37.990007025701459], [74.864815708316797, 38.378846340481601], [74.257514276022704, 38.60650686294349], [73.928852166646408, 38.505815334622724], [73.675379266254851, 39.431236884105573], [73.960013055318456, 39.660008449861721], [73.82224368682833, 39.893973497063143], [74.776862420556057, 40.366425279291633], [75.467827996730733, 40.562072251948678], [76.526368035797447, 40.427946071935139], [76.904484490877138, 41.066485907549662], [78.187196893226059, 41.185315863604814], [78.543660923175281, 41.58224254003872], [80.119430373051415, 42.123940741538235], [80.259990268885332, 42.349999294599087], [80.180150180994389, 42.920067857426858], [80.866206496101228, 43.180362046881015], [79.96610639844144, 44.917516994804629], [81.947070753918098, 45.317027492853157], [82.458925815769049, 45.539649563166506], [83.180483839860557, 47.330031236350749], [85.164290399113241, 47.000955715516113], [85.720483839870695, 47.452969468773091], [85.768232863308384, 48.455750637396903], [86.598776483103364, 48.549181626980612], [87.359970330762707, 49.214980780629162], [87.751264276076682, 49.297197984405472], [88.013832228551706, 48.599462795600601], [88.854297723346775, 48.069081732773014], [90.280825636763922, 47.693549099307916], [90.970809360724985, 46.888146063822944], [90.585768263718336, 45.719716091487498], [90.94553958533433, 45.28607330991025], [92.13389082231825, 45.115075995456436], [93.48073367714133, 44.975472113620015], [94.68892866412537, 44.352331854828464], [95.306875441471533, 44.241330878265472], [95.762454868556716, 43.319449164394626], [96.349395786527822, 42.725635280928657], [97.451757440178, 42.748889675460077], [99.515817498780024, 42.524691473961695], [100.845865513108293, 42.663804429691425], [101.833040399179964, 42.514872951826277], [103.312278273534815, 41.907468166667627], [104.522281935649033, 41.908346666016627], [104.96499393109346, 41.597409572916348], [106.129315627061686, 42.134327704428898], [107.744772576938004, 42.481515814781915], [109.243595819131457, 42.519446316084156], [110.412103306115313, 42.871233628911028], [111.129682244920247, 43.406834011400178], [111.829587843881399, 43.743118394539493], [111.66773725794323, 44.073175767587713], [111.348376906379457, 44.457441718110061], [111.873306105600278, 45.102079372735119], [112.43606245325887, 45.011645616224257], [113.463906691544224, 44.808893134127118], [114.460331658996068, 45.339816799493889], [115.985096470200148, 45.727235012386018], [116.717868280098884, 46.388202419615254], [117.42170128791426, 46.672732855814218], [118.87432579963874, 46.805412095723653], [119.663269891438773, 46.692679958678951], [119.772823927897576, 47.048058783550147], [118.86657433479499, 47.74706004494621], [118.064142694166748, 48.066730455103738], [117.295507440257467, 47.697709052107399], [116.308952671373248, 47.853410142602826], [115.742837355615762, 47.726544501326288], [115.485282017073047, 48.135382595403456], [116.191802199367629, 49.134598090199063], [116.678800897286209, 49.888531399121405], [117.879244419426385, 49.510983384796958], [119.288460728025854, 50.142882798862047], [119.279365675942387, 50.582907619827296], [120.182049595216967, 51.643566392618027], [120.738191359542014, 51.964115302124554], [120.725789015792003, 52.516226304730822], [120.177088657716894, 52.753886216841209], [121.003084751470254, 53.251401068731241], [122.245747918792887, 53.431725979213695], [123.571506789240885, 53.458804429734641], [125.068211297710462, 53.161044826868846], [125.946348911646197, 52.79279857035695], [126.564399041857001, 51.784255479532703], [126.939156528837685, 51.35389415140591], [127.287455682484932, 50.739797268265448], [127.657407261262421, 49.760270494172943]]]] } }
-    const geojson4 = { "_id": "4", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "Turkey", "sov_a3": "TUR", "adm0_dif": 0.0, "level": 2.0, "type": "Sovereign country", "admin": "Turkey", "adm0_a3": "TUR", "geou_dif": 0.0, "geounit": "Turkey", "gu_a3": "TUR", "su_dif": 0.0, "subunit": "Turkey", "su_a3": "TUR", "brk_diff": 0.0, "name": "Turkey", "name_long": "Turkey", "brk_a3": "TUR", "brk_name": "Turkey", "brk_group": null, "abbrev": "Tur.", "postal": "TR", "formal_en": "Republic of Turkey", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "Turkey", "name_alt": null, "mapcolor7": 6.0, "mapcolor8": 3.0, "mapcolor9": 8.0, "mapcolor13": 4.0, "pop_est": 76805524.0, "gdp_md_est": 902700.0, "pop_year": -99.0, "lastcensus": 2000.0, "gdp_year": -99.0, "economy": "4. Emerging region: MIKT", "income_grp": "3. Upper middle income", "wikipedia": -99.0, "fips_10": null, "iso_a2": "TR", "iso_a3": "TUR", "iso_n3": "792", "un_a3": "792", "wb_a2": "TR", "wb_a3": "TUR", "woe_id": -99.0, "adm0_a3_is": "TUR", "adm0_a3_us": "TUR", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "Asia", "region_un": "Asia", "subregion": "Western Asia", "region_wb": "Europe & Central Asia", "name_len": 6.0, "long_len": 6.0, "abbrev_len": 4.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "MultiPolygon", "coordinates": [[[[36.913127068842158, 41.335358384764305], [38.347664829264517, 40.948586127275718], [39.512606642420252, 41.102762763018575], [40.373432651538252, 41.013672593747344], [41.554084100110714, 41.535656236327611], [42.619548781104555, 41.583172715819927], [43.582745802592711, 41.092143256182567], [43.752657911968498, 40.740200914058818], [43.65643639504097, 40.253563951166171], [44.400008579288766, 40.005000311842309], [44.793989699082005, 39.713002631177034], [44.109225294782362, 39.428136298168056], [44.421402622257602, 38.281281236314527], [44.225755649600529, 37.971584377589352], [44.772699008977753, 37.170444647768448], [44.293451775902867, 37.00151439060636], [43.942258742047358, 37.256227525372935], [42.779125604021857, 37.385263576805812], [42.349591098811771, 37.229872544904111], [41.212089471203029, 37.074352321921737], [40.673259311695716, 37.091276353497364], [39.522580193852519, 36.716053778626019], [38.699891391765931, 36.712927354472328], [38.16772749202417, 36.901210435527787], [37.066761102045831, 36.623036200500621], [36.739494256341374, 36.817520453431115], [36.68538903173183, 36.259699205056506], [36.417550083163093, 36.040616970355103], [36.149762811026591, 35.821534735653671], [35.782084995269855, 36.274995429014922], [36.160821567537056, 36.650605577128374], [35.550936313628341, 36.565442816711339], [34.714553256984374, 36.795532131490916], [34.02689497247647, 36.219960028623973], [32.509158156064103, 36.1075637883892], [31.699595167779563, 36.644275214172609], [30.621624790171069, 36.677864895162315], [30.391096225717121, 36.26298065850699], [29.699975620245567, 36.144357408181008], [28.732902866335394, 36.676831366516438], [27.64118655773737, 36.658822129862756], [27.048767937943296, 37.653360907536012], [26.318218214633049, 38.208133246405396], [26.804700148228733, 38.985760199533559], [26.170785353304382, 39.463612168936464], [27.280019972449395, 40.420013739578309], [28.81997765474722, 40.460011298172219], [29.240003696415584, 41.219990749672689], [31.145933872204438, 41.087621568357065], [32.347979363745793, 41.736264146484643], [33.513282911927519, 42.018960069337311], [35.16770389175187, 42.040224921225445], [36.913127068842158, 41.335358384764305]]], [[[27.192376743282409, 40.690565700842455], [26.35800906749779, 40.151993923496491], [26.043351271272542, 40.617753607743168], [26.056942172965336, 40.824123440100749], [26.294602085075695, 40.936261298174173], [26.604195590936285, 41.56211456966102], [26.117041863720829, 41.826904608724561], [27.135739373490509, 42.141484890301314], [27.996720411905414, 42.007358710287775], [28.115524529744448, 41.622886054036286], [28.98844282401879, 41.299934190428189], [28.806438429486747, 41.054962063148537], [27.619017368284119, 40.999823309893117], [27.192376743282409, 40.690565700842455]]]] } }
-
-    const geojson3 = {
-        "_id": "3", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "Canada", "sov_a3": "CAN", "adm0_dif": 0.0, "level": 2.0, "type": "Sovereign country", "admin": "Canada", "adm0_a3": "CAN", "geou_dif": 0.0, "geounit": "Canada", "gu_a3": "CAN", "su_dif": 0.0, "subunit": "Canada", "su_a3": "CAN", "brk_diff": 0.0, "name": "Canada", "name_long": "Canada", "brk_a3": "CAN", "brk_name": "Canada", "brk_group": null, "abbrev": "Can.", "postal": "CA", "formal_en": "Canada", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "Canada", "name_alt": null, "mapcolor7": 6.0, "mapcolor8": 6.0, "mapcolor9": 2.0, "mapcolor13": 2.0, "pop_est": 33487208.0, "gdp_md_est": 1300000.0, "pop_year": -99.0, "lastcensus": 2011.0, "gdp_year": -99.0, "economy": "1. Developed region: G7", "income_grp": "1. High income: OECD", "wikipedia": -99.0, "fips_10": null, "iso_a2": "CA", "iso_a3": "CAN", "iso_n3": "124", "un_a3": "124", "wb_a2": "CA", "wb_a3": "CAN", "woe_id": -99.0, "adm0_a3_is": "CAN", "adm0_a3_us": "CAN", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "North America", "region_un": "Americas", "subregion": "Northern America", "region_wb": "North America", "name_len": 6.0, "long_len": 6.0, "abbrev_len": 4.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": {
-            "type": "MultiPolygon", "coordinates": [[[[-63.6645, 46.55001], [-62.9393, 46.41587], [-62.01208, 46.44314], [-62.50391, 46.03339], [-62.87433, 45.96818], [-64.1428, 46.39265], [-64.39261, 46.72747], [-64.01486, 47.03601], [-63.6645, 46.55001]]], [[[-61.8063, 49.10506], [-62.29318, 49.08717], [-63.58926, 49.40069], [-64.51912, 49.87304], [-64.17322, 49.95718], [-62.85829, 49.70641], [-61.835585, 49.28855], [-61.8063, 49.10506]]], [[[-123.510001587551159, 48.510010891303438], [-124.012890788399503, 48.370846259141416], [-125.655012777338371, 48.825004584338501], [-125.954994466792769, 49.179995835967645], [-126.850004435871881, 49.530000311880428], [-127.029993449544421, 49.81499583597008], [-128.059336304366241, 49.994959011426602], [-128.444584107102173, 50.539137681676124], [-128.358413656255436, 50.770648098343685], [-127.308581096029897, 50.552573554071955], [-126.695000977212317, 50.400903225295394], [-125.755006673823203, 50.295018215529382], [-125.415001587558805, 49.950000515332619], [-124.920768189119343, 49.475274970083404], [-123.922508708321018, 49.062483628935809], [-123.510001587551159, 48.510010891303438]]], [[[-56.134035814017125, 50.687009792679305], [-56.795881720595276, 49.81230866149096], [-56.143105027884303, 50.150117499382844], [-55.471492275602941, 49.935815334668462], [-55.822401089080927, 49.587128607779107], [-54.935142584845664, 49.31301097268684], [-54.473775397343786, 49.556691189159181], [-53.476549445191324, 49.249138902374057], [-53.78601375997124, 48.516780503933632], [-53.086133999226263, 48.687803656603535], [-52.958648240762244, 48.157164211614486], [-52.648098720904187, 47.535548407575504], [-53.069158291218343, 46.65549876564495], [-53.521456264853043, 46.61829173439483], [-54.178935512902541, 46.807065741557011], [-53.961868659060485, 47.625207017601923], [-54.240482143762137, 47.752279364607631], [-55.400773078011497, 46.884993801453135], [-55.997480841685842, 46.919720363953303], [-55.291219041552779, 47.389562486350997], [-56.250798712780522, 47.632545070987391], [-57.325229254777099, 47.572807115258001], [-59.266015184146767, 47.603347886742512], [-59.419494188053704, 47.899453843774864], [-58.796586473207412, 48.251525376979487], [-59.231624518456528, 48.523188381537807], [-58.391804979065228, 49.125580552764177], [-57.358689744686046, 50.718274034215852], [-56.738650071832012, 51.287438259478535], [-55.870976935435294, 51.632094224649194], [-55.406974249886616, 51.588272610065729], [-55.600218268442092, 51.317074693397927], [-56.134035814017125, 50.687009792679305]]], [[[-133.180004041711697, 54.169975490935315], [-132.71000788443132, 54.040009315423532], [-131.749989584003288, 54.120004380909222], [-132.049480347350993, 52.984621487024526], [-131.179042521826602, 52.180432847698285], [-131.577829549822923, 52.182370713909251], [-132.180428426778548, 52.639707139692405], [-132.549992432313871, 53.100014960332146], [-133.054611178755522, 53.411468817755377], [-133.239664482792705, 53.851080227262401], [-133.180004041711697, 54.169975490935315]]], [[[-79.26582, 62.158675], [-79.65752, 61.63308], [-80.09956, 61.7181], [-80.36215, 62.01649], [-80.31539, 62.085565], [-79.92939, 62.3856], [-79.52002, 62.36371], [-79.26582, 62.158675]]], [[[-81.89825, 62.7108], [-83.06857, 62.15922], [-83.77462, 62.18231], [-83.99367, 62.4528], [-83.25048, 62.91409], [-81.87699, 62.90458], [-81.89825, 62.7108]]], [[[-85.161307949549865, 65.657284654392811], [-84.975763719405961, 65.217518215588996], [-84.464012010419509, 65.371772365980178], [-83.882626308919754, 65.10961782496355], [-82.787576870438784, 64.766693020274687], [-81.642013719392537, 64.455135809986956], [-81.553440314444259, 63.979609280037153], [-80.817361212878865, 64.057485663501012], [-80.103451300766608, 63.725981350348604], [-80.991019863595682, 63.411246039474975], [-82.547178107417011, 63.651722317145243], [-83.108797573565056, 64.101875718839722], [-84.100416632813875, 63.569711819098018], [-85.523404710619019, 63.052379055424097], [-85.866768764982368, 63.637252916103563], [-87.221983201836736, 63.541238104905233], [-86.352759772471273, 64.035833238370714], [-86.224886440765147, 64.822916978608276], [-85.883847825854872, 65.738778388117055], [-85.161307949549865, 65.657284654392811]]], [[[-75.86588, 67.14886], [-76.98687, 67.09873], [-77.2364, 67.58809], [-76.81166, 68.14856], [-75.89521, 68.28721], [-75.1145, 68.01036], [-75.10333, 67.58202], [-75.21597, 67.44425], [-75.86588, 67.14886]]], [[[-95.647681203800516, 69.107690358321776], [-96.269521203800593, 68.757040358321746], [-97.617401203800569, 69.060030358321796], [-98.431801203800518, 68.950700358321797], [-99.797401203800533, 69.4000303583218], [-98.917401203800551, 69.710030358321802], [-98.218261203800495, 70.143540358321758], [-97.157401203800561, 69.860030358321808], [-96.557401203800538, 69.680030358321773], [-96.257401203800526, 69.490030358321775], [-95.647681203800516, 69.107690358321776]]], [[[-90.5471, 69.49766], [-90.55151, 68.47499], [-89.21515, 69.25873], [-88.01966, 68.61508], [-88.31749, 67.87338], [-87.35017, 67.19872], [-86.30607, 67.92146], [-85.57664, 68.78456], [-85.52197, 69.88211], [-84.10081, 69.80539], [-82.62258, 69.65826], [-81.28043, 69.16202], [-81.2202, 68.66567], [-81.96436, 68.13253], [-81.25928, 67.59716], [-81.38653, 67.11078], [-83.34456, 66.41154], [-84.73542, 66.2573], [-85.76943, 66.55833], [-86.0676, 66.05625], [-87.03143, 65.21297], [-87.32324, 64.77563], [-88.48296, 64.09897], [-89.91444, 64.03273], [-90.70398, 63.61017], [-90.77004, 62.96021], [-91.93342, 62.83508], [-93.15698, 62.02469], [-94.24153, 60.89865], [-94.62931, 60.11021], [-94.6846, 58.94882], [-93.21502, 58.78212], [-92.76462, 57.84571], [-92.29703, 57.08709], [-90.89769, 57.28468], [-89.03953, 56.85172], [-88.03978, 56.47162], [-87.32421, 55.99914], [-86.07121, 55.72383], [-85.01181, 55.3026], [-83.36055, 55.24489], [-82.27285, 55.14832], [-82.4362, 54.28227], [-82.12502, 53.27703], [-81.40075, 52.15788], [-79.91289, 51.20842], [-79.14301, 51.53393], [-78.60191, 52.56208], [-79.12421, 54.14145], [-79.82958, 54.66772], [-78.22874, 55.13645], [-77.0956, 55.83741], [-76.54137, 56.53423], [-76.62319, 57.20263], [-77.30226, 58.05209], [-78.51688, 58.80458], [-77.33676, 59.85261], [-77.77272, 60.75788], [-78.10687, 62.31964], [-77.41067, 62.55053], [-75.69621, 62.2784], [-74.6682, 62.18111], [-73.83988, 62.4438], [-72.90853, 62.10507], [-71.67708, 61.52535], [-71.37369, 61.13717], [-69.59042, 61.06141], [-69.62033, 60.22125], [-69.2879, 58.95736], [-68.37455, 58.80106], [-67.64976, 58.21206], [-66.20178, 58.76731], [-65.24517, 59.87071], [-64.58352, 60.33558], [-63.80475, 59.4426], [-62.50236, 58.16708], [-61.39655, 56.96745], [-61.79866, 56.33945], [-60.46853, 55.77548], [-59.56962, 55.20407], [-57.97508, 54.94549], [-57.3332, 54.6265], [-56.93689, 53.78032], [-56.15811, 53.64749], [-55.75632, 53.27036], [-55.68338, 52.14664], [-56.40916, 51.7707], [-57.12691, 51.41972], [-58.77482, 51.0643], [-60.03309, 50.24277], [-61.72366, 50.08046], [-63.86251, 50.29099], [-65.36331, 50.2982], [-66.39905, 50.22897], [-67.23631, 49.51156], [-68.51114, 49.06836], [-69.95362, 47.74488], [-71.10458, 46.82171], [-70.25522, 46.98606], [-68.65, 48.3], [-66.55243, 49.1331], [-65.05626, 49.23278], [-64.17099, 48.74248], [-65.11545, 48.07085], [-64.79854, 46.99297], [-64.47219, 46.23849], [-63.17329, 45.73902], [-61.52072, 45.88377], [-60.51815, 47.00793], [-60.4486, 46.28264], [-59.80287, 45.9204], [-61.03988, 45.26525], [-63.25471, 44.67014], [-64.24656, 44.26553], [-65.36406, 43.54523], [-66.1234, 43.61867], [-66.16173, 44.46512], [-64.42549, 45.29204], [-66.02605, 45.25931], [-67.13741, 45.13753], [-67.79134, 45.70281], [-67.79046, 47.06636], [-68.23444, 47.35486], [-68.905, 47.185], [-69.23722, 47.447781], [-69.99997, 46.69307], [-70.305, 45.915], [-70.66, 45.46], [-71.08482, 45.30524], [-71.405, 45.255], [-71.50506, 45.0082], [-73.34783, 45.00738], [-74.867, 45.00048], [-75.31821, 44.81645], [-76.375, 44.09631], [-76.5, 44.018458893758719], [-76.820034145805579, 43.628784288093755], [-77.737885097957701, 43.629055589363304], [-78.720279914042379, 43.625089423184875], [-79.171673550111876, 43.466339423184223], [-79.01, 43.27], [-78.92, 42.965], [-78.939362148743697, 42.863611355148038], [-80.247447679347943, 42.366199856122591], [-81.277746548167158, 42.209025987306859], [-82.439277716791622, 41.675105088867156], [-82.690089280920176, 41.675105088867156], [-83.029810146806938, 41.832795722005841], [-83.141999681312569, 41.975681057292832], [-83.12, 42.08], [-82.9, 42.43], [-82.43, 42.98], [-82.137642381503895, 43.571087551439916],
-            [-82.337763125431081, 44.44], [-82.550924648758183, 45.347516587905375], [-83.592850714843081, 45.81689362241238], [-83.469550747394635, 45.994686387712591], [-83.616130947590591, 46.11692698829907], [-83.890765347005754, 46.11692698829907], [-84.091851264161477, 46.275418606138174], [-84.142119513673379, 46.512225857115737], [-84.3367, 46.40877], [-84.6049, 46.4396], [-84.543748745445868, 46.538684190449146], [-84.779238247399917, 46.637101955749046], [-84.876079881514869, 46.900083319682381], [-85.652363247403429, 47.220218817730512], [-86.461990831228263, 47.553338019392044], [-87.439792623300235, 47.94], [-88.378114183286726, 48.302917588893735], [-89.272917446636683, 48.019808254582671], [-89.6, 48.01], [-90.83, 48.27], [-91.64, 48.14], [-92.61, 48.45], [-93.63087, 48.60926], [-94.32914, 48.67074], [-94.64, 48.84], [-94.81758, 49.38905], [-95.15609, 49.38425], [-95.159069509172042, 49.0], [-97.228720000004813, 49.0007], [-100.65, 49.0], [-104.04826, 48.99986], [-107.05, 49.0], [-110.05, 49.0], [-113.0, 49.0], [-116.04818, 49.0], [-117.03121, 49.0], [-120.0, 49.0], [-122.84, 49.0], [-122.97421, 49.002537777777803], [-124.91024, 49.98456], [-125.62461, 50.41656], [-127.43561, 50.83061], [-127.99276, 51.71583], [-127.85032, 52.32961], [-129.12979, 52.75538], [-129.30523, 53.56159], [-130.51497, 54.28757], [-130.53611, 54.80278], [-129.98, 55.285], [-130.00778, 55.91583], [-131.70781, 56.55212], [-132.73042, 57.69289], [-133.35556, 58.41028], [-134.27111, 58.86111], [-134.945, 59.27056], [-135.47583, 59.78778], [-136.47972, 59.46389], [-137.4525, 58.905], [-138.34089, 59.56211], [-139.039, 60.0], [-140.013, 60.27682], [-140.99778, 60.30639], [-140.9925, 66.00003], [-140.986, 69.712], [-139.12052, 69.47102], [-137.54636, 68.99002], [-136.50358, 68.89804], [-135.62576, 69.31512], [-134.41464, 69.62743], [-132.92925, 69.50534], [-131.43136, 69.94451], [-129.79471, 70.19369], [-129.10773, 69.77927], [-128.36156, 70.01286], [-128.13817, 70.48384], [-127.44712, 70.37721], [-125.75632, 69.48058], [-124.42483, 70.1584], [-124.28968, 69.39969], [-123.06108, 69.56372], [-122.6835, 69.85553], [-121.47226, 69.79778], [-119.94288, 69.37786], [-117.60268, 69.01128], [-116.22643, 68.84151], [-115.2469, 68.90591], [-113.89794, 68.3989], [-115.30489, 67.90261], [-113.49727, 67.68815], [-110.798, 67.80612], [-109.94619, 67.98104], [-108.8802, 67.38144], [-107.79239, 67.88736], [-108.81299, 68.31164], [-108.16721, 68.65392], [-106.95, 68.7], [-106.15, 68.8], [-105.34282, 68.56122], [-104.33791, 68.018], [-103.22115, 68.09775], [-101.45433, 67.64689], [-99.90195, 67.80566], [-98.4432, 67.78165], [-98.5586, 68.40394], [-97.66948, 68.57864], [-96.11991, 68.23939], [-96.12588, 67.29338], [-95.48943, 68.0907], [-94.685, 68.06383], [-94.23282, 69.06903], [-95.30408, 69.68571], [-96.47131, 70.08976], [-96.39115, 71.19482], [-95.2088, 71.92053], [-93.88997, 71.76015], [-92.87818, 71.31869], [-91.51964, 70.19129], [-92.40692, 69.69997], [-90.5471, 69.49766]]], [[[-114.167169999999899, 73.12145], [-114.66634, 72.65277], [-112.441019999999895, 72.955400000000111], [-111.05039, 72.4504], [-109.920349999999885, 72.961130000000111], [-109.00654, 72.63335], [-108.18835, 71.65089], [-107.68599, 72.06548], [-108.39639, 73.08953000000011], [-107.51645, 73.23598], [-106.52259, 73.07601], [-105.40246, 72.67259], [-104.77484, 71.698400000000106], [-104.464759999999842, 70.99297], [-102.78537, 70.49776], [-100.980779999999896, 70.02432], [-101.08929, 69.584470000000124], [-102.73116, 69.50402], [-102.09329, 69.119620000000111], [-102.43024, 68.75282], [-104.24, 68.91], [-105.96, 69.180000000000149], [-107.12254, 69.11922], [-109.0, 68.78], [-111.534148875200145, 68.630059156817936], [-113.3132, 68.53554], [-113.854959999999835, 69.007440000000116], [-115.22, 69.28], [-116.10794, 69.16821], [-117.34, 69.960000000000122], [-116.674729999999897, 70.06655], [-115.13112, 70.2373], [-113.72141, 70.19237], [-112.4161, 70.36638], [-114.35, 70.6], [-116.48684, 70.52045], [-117.9048, 70.540560000000141], [-118.43238, 70.9092], [-116.11311, 71.30918], [-117.65568, 71.2952], [-119.40199, 71.55859], [-118.56267, 72.30785], [-117.86642, 72.70594], [-115.18909, 73.314590000000123], [-114.167169999999899, 73.12145]]], [[[-104.5, 73.42], [-105.38, 72.76], [-106.94, 73.46], [-106.6, 73.6], [-105.26, 73.64], [-104.5, 73.42]]], [[[-76.34, 73.102684989953019], [-76.25140380859375, 72.826385498046875], [-77.314437866210909, 72.855545043945284], [-78.39167022705081, 72.876655578613281], [-79.486251831054659, 72.742202758789091], [-79.775833129882841, 72.802902221679744], [-80.876098632812528, 73.333183288574219], [-80.833885192871065, 73.693183898925781], [-80.353057861328125, 73.759719848632784], [-78.064437866210938, 73.651931762695341], [-76.34, 73.102684989953019]]], [[[-86.562178514334136, 73.157447007938458], [-85.774371304044536, 72.534125881633827], [-84.850112474288238, 73.340278225387124], [-82.315590176100983, 73.750950832810588], [-80.600087653307639, 72.71654368762421], [-80.748941616524405, 72.061906643350767], [-78.770638597310779, 72.352173163534161], [-77.824623989559598, 72.749616604291049], [-75.605844692675731, 72.243678493937409], [-74.228616095664989, 71.767144273557903], [-74.099140794557712, 71.330840155717652], [-72.242225714797655, 71.556924546994509], [-71.200015428335206, 70.920012518997225], [-68.786054246684898, 70.525023708774256], [-67.914970465756937, 70.121947536897608], [-66.969033372654167, 69.186087348091888], [-68.805122850200547, 68.720198472764423], [-66.449866095633865, 68.067163397892017], [-64.862314419195229, 67.847538560651628], [-63.424934454996759, 66.928473212340663], [-61.851981370680591, 66.862120673277843], [-62.163176845942303, 66.160251369889608], [-63.918444383384184, 64.998668524832851], [-65.148860236253626, 65.426032619886684], [-66.721219041598545, 66.388041083432199], [-68.015016038673963, 66.262725735124405], [-68.141287400979166, 65.689789130304376], [-67.089646165623407, 65.108455105236999], [-65.732080451099762, 64.648405666758634], [-65.320167609301279, 64.382737128346065], [-64.669406297449683, 63.392926744227481], [-65.013803880458909, 62.674185085695996], [-66.275044725190469, 62.945098781986076], [-68.783186204692726, 63.745670071051819], [-67.369680752213043, 62.883965562584876], [-66.328297288667216, 62.280074774822054], [-66.165568203380161, 61.930897121825893], [-68.877366502544646, 62.330149237712817], [-71.023437059193839, 62.910708116295837], [-72.235378587518994, 63.397836005295176], [-71.8862784491713, 63.679989325608858], [-73.378306240518384, 64.193963121183828], [-74.834418911422603, 64.679075629323791], [-74.818502570276735, 64.389093329517976], [-77.709979824520047, 64.229542344816792], [-78.555948859354174, 64.572906399180141], [-77.897281053361922, 65.309192206474791], [-76.018274298797195, 65.326968899183157], [-73.959795294882724, 65.454764716240902], [-74.29388342964964, 65.811771348729394], [-73.944912482382648, 66.310578111426736], [-72.651167161739409, 67.284575507263867], [-72.926059943316091, 67.726925767682388], [-73.311617804645749, 68.069437160912912], [-74.843307257776814, 68.554627183701285], [-76.869100918266753, 68.894735622830268], [-76.228649054657353, 69.147769273547425], [-77.287369961237118, 69.769540106883284], [-78.168633999326602, 69.82648753526891], [-78.957242194316734, 70.166880194775416], [-79.492455003563663, 69.871807766388912], [-81.30547095409176, 69.743185126414346], [-84.94470618359847, 69.966634019644403], [-87.060003424817893, 70.26000112576537], [-88.681713223001509, 70.41074127876081], [-89.51341956252304, 70.762037665480989], [-88.467721116880767, 71.218185533321332], [-89.888151211287493, 71.222552191849957], [-90.205160285182018, 72.235074367960806], [-89.436576707704944, 73.129464219852366], [-88.408241543312812, 73.537888902471224], [-85.82615108920092, 73.803815823045227], [-86.562178514334136, 73.157447007938458]]], [[[-100.35642, 73.84389], [-99.16387, 73.63339], [-97.38, 73.76], [-97.12, 73.47], [-98.05359, 72.99052], [-96.54, 72.56], [-96.72, 71.66], [-98.35966, 71.27285], [-99.32286, 71.35639], [-100.01482, 71.73827], [-102.5, 72.51], [-102.48, 72.83], [-100.43836, 72.70588], [-101.54, 73.36], [-100.35642, 73.84389]]], [[[-93.19629553910022, 72.771992499473356], [-94.269046597047264, 72.024596259235977], [-95.409855516322665, 72.061880805134592], [-96.033745083382456, 72.940276801231818], [-96.018267991910989, 73.437429918095802], [-95.495793423224029, 73.862416897264183], [-94.503657599652342, 74.13490672473921], [-92.420012173211774, 74.100025132942193], [-90.509792853542592, 73.856732489712044], [-92.003965216829897, 72.966244208458505], [-93.19629553910022, 72.771992499473356]]], [[[-120.46, 71.383601793087593], [-123.09219, 70.90164], [-123.62, 71.34], [-125.928948737473334, 71.868688463011409], [-125.5, 72.292260811795018], [-124.80729, 73.02256], [-123.939999999999898, 73.680000000000149], [-124.91775, 74.292750000000126], [-121.53788, 74.44893], [-120.10978, 74.24135], [-117.555639999999869, 74.18577], [-116.58442, 73.89607], [-115.51081, 73.47519], [-116.767939999999896, 73.22292], [-119.22, 72.52], [-120.46, 71.82], [-120.46, 71.383601793087593]]], [[[-93.612755906940492, 74.979997260224451], [-94.15690873897384, 74.592346503386864], [-95.608680589565608, 74.666863918751773], [-96.820932176484575, 74.92762319609659], [-96.28858740922982, 75.377828274223361], [-94.850819871789128, 75.6472175157609], [-93.977746548217937, 75.296489569795966], [-93.612755906940492, 74.979997260224451]]], [[[-98.5, 76.72], [-97.73558, 76.25656], [-97.704415, 75.74344], [-98.16, 75.0], [-99.80874, 74.89744], [-100.88366, 75.05736], [-100.86292, 75.64075], [-102.50209, 75.5638], [-102.56552, 76.3366], [-101.48973, 76.30537], [-99.98349, 76.64634], [-98.57699, 76.58859], [-98.5, 76.72]]], [[[-108.21141, 76.20168], [-107.81943, 75.84552], [-106.92893, 76.01282], [-105.881, 75.9694], [-105.70498, 75.47951], [-106.31347, 75.00527], [-109.7, 74.85], [-112.22307, 74.41696], [-113.74381, 74.39427], [-113.87135, 74.72029], [-111.79421, 75.1625], [-116.31221, 75.04343], [-117.7104, 75.2222], [-116.34602, 76.19903], [-115.40487, 76.47887], [-112.59056, 76.14134], [-110.81422, 75.54919], [-109.0671, 75.47321], [-110.49726, 76.42982], [-109.5811, 76.79417], [-108.54859, 76.67832], [-108.21141, 76.20168]]], [[[-94.684085862999467, 77.097878323058382], [-93.573921068073133, 76.776295884906091], [-91.605023159536614, 76.778517971494608], [-90.741845872749224, 76.449597479956822], [-90.96966142450799, 76.074013170059459], [-89.822237921899273, 75.84777374948564], [-89.18708289259979, 75.61016551380763], [-87.838276333349626, 75.566188869927231], [-86.379192267588678, 75.482421373182177], [-84.789625210290609, 75.699204006646511], [-82.753444586910064, 75.784315090631253], [-81.128530849924374, 75.713983466282031], [-80.057510952459154, 75.336848863415895], [-79.833932868148338, 74.923127346487206], [-80.457770758775837, 74.657303778777788], [-81.948842536125539, 74.442459011524335], [-83.228893602211429, 74.564027818490956], [-86.097452358733307, 74.410032050261151], [-88.150350307960224, 74.392307033984991], [-89.764722052758373, 74.515555325001145], [-92.422440965529432, 74.837757880341002], [-92.768285488642817, 75.386819973442158], [-92.889905972041731, 75.882655341282657], [-93.893824022176005, 76.319243679500545], [-95.962457445035824, 76.441380927222468], [-97.121378953829492, 76.751077785947615], [-96.745122850312356, 77.161388658345146], [-94.684085862999467, 77.097878323058382]]], [[[-116.198586595507351, 77.645286770326209], [-116.335813361458392, 76.876961575010569], [-117.106050584768795, 76.530031846819128], [-118.040412157038148, 76.481171780087095], [-119.899317586885701, 76.053213406061985], [-121.499995077126499, 75.900018622532798], [-122.854924486158978, 76.116542873835698], [-122.854925293603216, 76.116542873835698], [-121.157535360328254, 76.86450755482835], [-119.103938971821051, 77.512219957174636], [-117.570130784965968, 77.498318996888116], [-116.198586595507351, 77.645286770326209]]], [[[-93.840003017943985, 77.519997260234504], [-94.295608283245258, 77.491342678528696], [-96.169654100310083, 77.555111395976894], [-96.436304490936124, 77.834629218243634], [-94.422577277386381, 77.820004787904992], [-93.720656297565881, 77.634331366680328], [-93.840003017943985, 77.519997260234504]]], [[[-110.186938035912974, 77.6970148790503], [-112.051191169058484, 77.409228827616857], [-113.534278937619064, 77.732206529441157], [-112.724586758253849, 78.051050116681949], [-111.26444332563085, 78.152956041161559], [-109.854451870547095, 77.99632477488484], [-110.186938035912974, 77.6970148790503]]], [[[-109.663145718202585, 78.60197256134569], [-110.881314256618865, 78.406919867660008], [-112.54209143761517, 78.407901719873507], [-112.525890876091594, 78.550554511215239], [-111.500010342233395, 78.849993598130567], [-110.963660651476019, 78.804440823065221], [-109.663145718202585, 78.60197256134569]]], [[[-95.830294969449341, 78.056941229963257], [-97.309842902398003, 77.850597235821795], [-98.124289313533978, 78.082856960757596], [-98.552867804746654, 78.458105373845115], [-98.631984422585532, 78.871930243638388], [-97.337231411512619, 78.831984361476771], [-96.754398769908789, 78.765812689927003], [-95.559277920294591, 78.418314520980289], [-95.830294969449341, 78.056941229963257]]], [[[-100.06019182005214, 78.324754340315906], [-99.670939093813615, 77.907544664207421], [-101.303940192453013, 78.018984890444813], [-102.949808722733053, 78.34322866486022], [-105.176132778731542, 78.380332343245755], [-104.210429450277161, 78.677420152491806], [-105.41958045125854, 78.918335679836446], [-105.492289191493157, 79.301593939929205], [-103.529282396237932, 79.165349026191649], [-100.825158047268815, 78.800461737778704], [-100.06019182005214, 78.324754340315906]]], [[[-87.02, 79.66], [-85.81435, 79.3369], [-87.18756, 79.0393], [-89.03535, 78.28723], [-90.80436, 78.21533], [-92.87669, 78.34333], [-93.95116, 78.75099], [-93.93574, 79.11373], [-93.14524, 79.3801], [-94.974, 79.37248], [-96.07614, 79.70502], [-96.70972, 80.15777], [-96.01644, 80.60233], [-95.32345, 80.90729], [-94.29843, 80.97727], [-94.73542, 81.20646], [-92.40984, 81.25739], [-91.13289, 80.72345], [-89.45, 80.509322033898286], [-87.81, 80.32], [-87.02, 79.66]]], [[[-68.5, 83.106321516765746], [-65.82735, 83.02801], [-63.68, 82.9], [-61.85, 82.6286], [-61.89388, 82.36165], [-64.334, 81.92775], [-66.75342, 81.72527], [-67.65755, 81.50141], [-65.48031, 81.50657], [-67.84, 80.9], [-69.4697, 80.61683], [-71.18, 79.8], [-73.2428, 79.63415], [-73.88, 79.430162204802087], [-76.90773, 79.32309], [-75.52924, 79.19766], [-76.22046, 79.01907], [-75.39345, 78.52581], [-76.34354, 78.18296], [-77.88851, 77.89991], [-78.36269, 77.50859], [-79.75951, 77.20968], [-79.61965, 76.98336], [-77.91089, 77.022045], [-77.88911, 76.777955], [-80.56125, 76.17812], [-83.17439, 76.45403], [-86.11184, 76.29901], [-87.6, 76.42], [-89.49068, 76.47239], [-89.6161, 76.95213], [-87.76739, 77.17833], [-88.26, 77.9], [-87.65, 77.970222222222219], [-84.97634, 77.53873], [-86.34, 78.18], [-87.96192, 78.37181], [-87.15198, 78.75867], [-85.37868, 78.9969], [-85.09495, 79.34543], [-86.50734, 79.73624], [-86.93179, 80.25145], [-84.19844, 80.20836], [-83.408695652173833, 80.1], [-81.84823, 80.46442], [-84.1, 80.58], [-87.59895, 80.51627], [-89.36663, 80.85569], [-90.2, 81.26], [-91.36786, 81.5531], [-91.58702, 81.89429], [-90.1, 82.085], [-88.93227, 82.11751], [-86.97024, 82.27961], [-85.5, 82.652273458057039], [-84.26, 82.6], [-83.18, 82.32], [-82.42, 82.86], [-81.1, 83.02], [-79.30664, 83.13056], [-76.25, 83.172058823529397], [-75.71878, 83.06404], [-72.83153, 83.23324], [-70.66576, 83.169780758382842], [-68.5, 83.106321516765746]]]]
-        }
-    }
-
-    const geojson6 = { "_id": "6", "type": "Feature", "properties": { "scalerank": 1, "featurecla": "Admin-0 country", "labelrank": 2.0, "sovereignt": "Brazil", "sov_a3": "BRA", "adm0_dif": 0.0, "level": 2.0, "type": "Sovereign country", "admin": "Brazil", "adm0_a3": "BRA", "geou_dif": 0.0, "geounit": "Brazil", "gu_a3": "BRA", "su_dif": 0.0, "subunit": "Brazil", "su_a3": "BRA", "brk_diff": 0.0, "name": "Brazil", "name_long": "Brazil", "brk_a3": "BRA", "brk_name": "Brazil", "brk_group": null, "abbrev": "Brazil", "postal": "BR", "formal_en": "Federative Republic of Brazil", "formal_fr": null, "note_adm0": null, "note_brk": null, "name_sort": "Brazil", "name_alt": null, "mapcolor7": 5.0, "mapcolor8": 6.0, "mapcolor9": 5.0, "mapcolor13": 7.0, "pop_est": 198739269.0, "gdp_md_est": 1993000.0, "pop_year": -99.0, "lastcensus": 2010.0, "gdp_year": -99.0, "economy": "3. Emerging region: BRIC", "income_grp": "3. Upper middle income", "wikipedia": -99.0, "fips_10": null, "iso_a2": "BR", "iso_a3": "BRA", "iso_n3": "076", "un_a3": "076", "wb_a2": "BR", "wb_a3": "BRA", "woe_id": -99.0, "adm0_a3_is": "BRA", "adm0_a3_us": "BRA", "adm0_a3_un": -99.0, "adm0_a3_wb": -99.0, "continent": "South America", "region_un": "Americas", "subregion": "South America", "region_wb": "Latin America & Caribbean", "name_len": 6.0, "long_len": 6.0, "abbrev_len": 6.0, "tiny": -99.0, "homepart": 1.0 }, "geometry": { "type": "Polygon", "coordinates": [[[-57.625133429582959, -30.216294854454262], [-56.290899624239081, -28.852760512000895], [-55.162286342984572, -27.881915378533463], [-54.490725267135524, -27.474756768505792], [-53.648735317587892, -26.92347258881609], [-53.628348965048744, -26.124865004177472], [-54.13004960795439, -25.547639255477254], [-54.625290696823576, -25.739255466415514], [-54.428946092330591, -25.162184747012166], [-54.293476325077449, -24.570799655863965], [-54.292959560754518, -24.021014092710729], [-54.652834235235133, -23.839578138933959], [-55.02790178080955, -24.001273695575229], [-55.400747239795422, -23.956935316668805], [-55.517639329639636, -23.571997572526637], [-55.610682745981144, -22.655619398694846], [-55.797958136606908, -22.356929620047822], [-56.473317430229393, -22.086300144135283], [-56.881509568902899, -22.282153822521479], [-57.937155727761294, -22.090175876557172], [-57.8706739976178, -20.732687676681952], [-58.166392381408045, -20.176700941653678], [-57.853801642474508, -19.969995212486189], [-57.949997321185826, -19.400004164306822], [-57.676008877174311, -18.961839694904029], [-57.498371141170992, -18.174187513911292], [-57.734558274961003, -17.552468357007768], [-58.280804002502251, -17.271710300366017], [-58.388058437724041, -16.877109063385276], [-58.24121985536668, -16.299573256091293], [-60.158389655179036, -16.258283786690086], [-60.542965664295153, -15.093910414289596], [-60.251148851142936, -15.07721892665932], [-60.26432634137737, -14.645979099183641], [-60.459198167550028, -14.354007256734555], [-60.503304002511136, -13.775954685117659], [-61.084121263255653, -13.479383640194598], [-61.713204311760776, -13.489202162330052], [-62.127080857986385, -13.198780612849724], [-62.803060268796386, -13.000653171442686], [-63.196498786050569, -12.627032565972435], [-64.316352912031604, -12.461978041232193], [-65.402281460213032, -11.566270440317155], [-65.321898769783019, -10.895872084194679], [-65.444837002205389, -10.511451104375432], [-65.338435228116424, -9.761987806846392], [-66.646908331962806, -9.931331475466862], [-67.173801235610739, -10.306812432499612], [-68.048192308205387, -10.712059014532485], [-68.271253628193264, -11.014521172736821], [-68.786157599549483, -11.03638030359628], [-69.529678107364958, -10.951734307502194], [-70.093752204046893, -11.123971856331012], [-70.548685675728407, -11.009146823778465], [-70.481893886991173, -9.490118096558845], [-71.302412278921537, -10.079436130415374], [-72.18489071316985, -10.053597914269432], [-72.563033006465645, -9.520193780152717], [-73.226713426390162, -9.462212823121234], [-73.015382656532552, -9.032833347208062], [-73.571059332967067, -8.424446709835834], [-73.987235480429661, -7.523829847853065], [-73.723401455363501, -7.340998630404414], [-73.724486660441642, -6.91859547285064], [-73.120027431923603, -6.629930922068239], [-73.21971126981461, -6.089188734566078], [-72.9645072089412, -5.741251315944893], [-72.891927659787257, -5.274561455916981], [-71.748405727816547, -4.593982842633011], [-70.928843349883579, -4.401591485210368], [-70.794768846302304, -4.251264743673303], [-69.893635219996625, -4.298186944194327], [-69.444101935489613, -1.556287123219818], [-69.420485805932231, -1.122618503426409], [-69.5770653957766, -0.549991957200163], [-70.020655890570055, -0.185156345219539], [-70.015565761989308, 0.541414292804205], [-69.452396002872462, 0.706158758950693], [-69.252434048119056, 0.602650865070075], [-69.21863766140018, 0.985676581217433], [-69.80459672715773, 1.089081122233466], [-69.816973232691623, 1.714805202639624], [-67.868565029558837, 1.692455145673392], [-67.537810024674698, 2.03716278727633], [-67.259997524673594, 1.719998684084956], [-67.065048183852497, 1.130112209473225], [-66.87632585312258, 1.253360500489336], [-66.325765143484958, 0.724452215982012], [-65.548267381437569, 0.78925446207603], [-65.354713304288367, 1.0952822941085], [-64.611011928959869, 1.328730576987042], [-64.199305792890513, 1.49285492594602], [-64.083085496666087, 1.91636912679408], [-63.368788011311665, 2.200899562993129], [-63.422867397705119, 2.411067613124175], [-64.269999152265797, 2.497005520025567], [-64.408827887617917, 3.126786200366624], [-64.368494432214106, 3.797210394705246], [-64.816064012294021, 4.056445217297423], [-64.628659430587547, 4.14848094320925], [-63.888342861574159, 4.020530096854571], [-63.093197597899106, 3.770571193858785], [-62.804533047116706, 4.006965033377952], [-62.085429653559132, 4.162123521334308], [-60.966893276601539, 4.536467596856639], [-60.601179165271944, 4.91809804933213], [-60.733574184803722, 5.200277207861901], [-60.213683437731333, 5.244486395687602], [-59.980958624904886, 5.014061184098139], [-60.11100236676738, 4.574966538914083], [-59.767405768458715, 4.423502915866607], [-59.538039923731233, 3.958802598481938], [-59.815413174057866, 3.606498521332085], [-59.974524909084558, 2.755232652188056], [-59.718545701726747, 2.24963043864436], [-59.646043667221257, 1.786893825686789], [-59.030861579002647, 1.317697658692722], [-58.540012986878295, 1.268088283692521], [-58.429477098205965, 1.463941962078721], [-58.113449876525017, 1.507195135907025], [-57.660971035377372, 1.682584947105639], [-57.335822923396904, 1.948537705895759], [-56.782704230360828, 1.863710842288654], [-56.539385748914555, 1.899522609866921], [-55.995698004771754, 1.817667141116601], [-55.905600145070885, 2.02199575439866], [-56.073341844290297, 2.220794989425499], [-55.973322109589375, 2.510363877773017], [-55.569755011605999, 2.421506252447131], [-55.097587449755139, 2.523748073736613], [-54.524754197799716, 2.311848863123785], [-54.08806250671725, 2.105556545414629], [-53.778520677288917, 2.376702785650082], [-53.554839240113544, 2.334896551925951], [-53.418465135295307, 2.053389187015981], [-52.939657151894956, 2.124857692875636], [-52.556424730018421, 2.504705308437053], [-52.249337531123956, 3.241094468596245], [-51.657797410678889, 4.156232408053029], [-51.317146369010857, 4.203490505383954], [-51.069771287629656, 3.650397650564031], [-50.508875291533656, 1.901563828942457], [-49.974075893745059, 1.736483465986069], [-49.947100796088712, 1.046189683431223], [-50.699251268096916, 0.222984117021682], [-50.388210822132137, -0.078444512536819], [-48.62056677915632, -0.235489190271821], [-48.584496629416591, -1.237805271005001], [-47.824956427590635, -0.5816179337628], [-46.566583624851226, -0.941027520352776], [-44.905703090990414, -1.551739597178134], [-44.417619187993665, -2.137750339367976], [-44.581588507655781, -2.691308282078524], [-43.418791266440195, -2.383110039889793], [-41.472656826328247, -2.912018324397116], [-39.978665330554037, -2.873054294449041], [-38.500383470196567, -3.700652357603396], [-37.2232521225352, -4.820945733258917], [-36.452937384576387, -5.109403578312154], [-35.597795783010469, -5.149504489770649], [-35.23538896334756, -5.464937432480247], [-34.896029832486832, -6.738193047719711], [-34.729993455533034, -7.343220716992967], [-35.128212042774223, -8.996401462442286], [-35.636966518687714, -9.649281508017815], [-37.046518724096998, -11.040721123908803], [-37.683611619607362, -12.171194756725823], [-38.423876512188443, -13.038118584854288], [-38.673887091616521, -13.057652276260619], [-38.953275722802545, -13.793369642800023], [-38.882298143049653, -15.667053724838768], [-39.161092495264313, -17.208406670808472], [-39.267339240056401, -17.867746270420483], [-39.583521491034233, -18.262295830968938], [-39.760823330227637, -19.599113457927409], [-40.774740770010339, -20.904511814052423], [-40.944756232250612, -21.937316989837811], [-41.754164191238225, -22.370675551037458], [-41.98828426773656, -22.970070489190896], [-43.074703742024752, -22.967693373305469], [-44.647811855637812, -23.351959323827842], [-45.352135789559917, -23.796841729428582], [-46.472093268405537, -24.088968601174543], [-47.648972337420659, -24.885199069927722], [-48.495458136577703, -25.877024834905654], [-48.64100480812774, -26.623697605090932], [-48.474735887228654, -27.175911960561891], [-48.661520351747626, -28.18613453543572], [-48.8884574041574, -28.674115085567884], [-49.587329474472675, -29.224469089476337], [-50.696874152211485, -30.98446502047296], [-51.576226162306156, -31.777698256153212], [-52.256081305538046, -32.24536996839467], [-52.712099982297694, -33.196578057591182], [-53.373661668498244, -33.768377780900764], [-53.650543992718099, -33.20200408298183], [-53.209588995971544, -32.727666110974724], [-53.787951626182192, -32.047242526987624], [-54.572451544805119, -31.494511407193748], [-55.601510179249345, -30.853878676071393], [-55.973244594940937, -30.883075860316303], [-56.976025763564735, -30.109686374636127], [-57.625133429582959, -30.216294854454262]]] } }
-
-    // handleLoad = map => {
-    //     const MapboxLanguage = require('@mapbox/mapbox-gl-language');
-    //     const language = new MapboxLanguage();
-    //     map.addControl(language);
-
-    //     // force call style update
-    //     language._initialStyleUpdate();
-    // }
-
-    // const changeMapLanguage = (map) => {
-    //     map.getStyle().layers.forEach((layer) => {
-    //         if (layer.id.endsWith('-label')) {
-    //             map.setLayoutProperty(layer.id, 'text-field', [
-    //                 'coalesce',
-    //                 ['get', 'name_ru'],
-    //                 ['get', 'name_ru'],
-    //             ]);
-    //         }
-    //     });
-    // };
-
-
-    // const changeMapLanguage = (map) => {
-    //     map.setLayoutProperty('country-label', 'text-field', [
-    //         'get',
-    //         'name_' + selectedLanguage
-    //     ]);
-    // }
-
-    // map.getStyle().layers.forEach(function(thisLayer){
-    //     console.log(thisLayer);
-    //         if(thisLayer.id.indexOf('-label')>0){
-    //             console.log('change '+thisLayer.id);
-    //             map.setLayoutProperty(thisLayer.id, 'text-field', ['get','name_fr'])
-    //         }
-    //     });
-
-    const changeMapLanguage = (map) => {
-        console.log(map)
-        // map.setLayoutProperty('country-label', 'text-field', [
-        //     'get',
-        // //     'name_ru'
-        // ]);
-        // map.getStyle().layers.forEach((layer) => {
-        //     if (layer.id.endsWith('-label')) {
-        //         map.setLayoutProperty(layer.id, 'text-field', [
-        //             'coalesce',
-        //             ['get', 'name_ru'],
-        //             ['get', 'name'],
-        //         ]);
-        //     }
-        // });
-    };
-
-    return (
-        <div>
-            <ReactMapboxGL
-
-
-                style={{ width: '100%', height: '445px' }}
-                mapStyle="mapbox://styles/general-m/ckmbenfqo4bxi17qivmeeyxo9"
-                accessToken={MAPBOX_ACCESS_TOKEN}
-                latitude={viewport.latitude}
-                longitude={viewport.longitude}
-                zoom={viewport.zoom}
-                onViewportChange={setViewport}
-                onLoad={changeMapLanguage}
-
-
-            >
-                <Marker latitude={latitudeCapital} longitude={longitudeCapital}>
-                    <div className={classes.marker}>
-                    </div>
-
-                </Marker>
-                <FullscreenControl position='top-right' />
-                <Source id={selectedCountryIndex.toString()} type="geojson" data={polygon}>
-                    <Layer
-                        id='anything'
-                        type='fill'
-                        paint={{
-                            'fill-color': '#f4cb67',
-                            'fill-opacity': 0.6
-                        }}
-                        source={selectedCountryIndex.toString()} />
-                </Source>
-            </ReactMapboxGL>
-
-        </div >
-    )
-}
-
-const mapStateToProps = (state: RootStateType) => {
-    return state.countryState;
-};
-
-
-export default connect(mapStateToProps, actions)(MapComponent);
-
-
-
-
-
-
-
-
-
