+++ travel-app/src/react-app-env.d.ts
@@ -3,3 +3,4 @@ declare module '*.mp4' {
   const src: string;
   export default src;
 }
+declare module 'urbica/react-map-gl'
